+++ packages/core/src/extensions/withConnectHook.ts
@@ -1,14 +1,14 @@
 import type { AtomLike, Ext } from '../core'
 import { action } from '../core'
 import { abortVar, wrap } from '../methods'
-import type { Unsubscribe } from '../utils'
+import type { Fn, MaybeUnsubscribe } from '../utils'
 import { isAbort, noop } from '../utils'
 import { withAbort } from './withAbort'
 import { withCallHook } from './withChangeHook'
 
 export let withConnectHook =
   <Target extends AtomLike>(
-    cb: (target: Target) => void | Unsubscribe | Promise<void | Unsubscribe>,
+    cb: (target: Target) => MaybeUnsubscribe,
   ): Ext<Target> =>
   (target) => {
     const onConnect = (target.__reatom.onConnect ??=